refactor: move post schema to schemas/content/ and rename PostSchema->PostResourceSchema
