+++ pnpm-workspace.yaml
@@ -1,0 +1,18 @@
+allowBuilds:
+  '@prisma/client': set this to true or false
+  '@prisma/engines': set this to true or false
+  aws-sdk: set this to true or false
+  bufferutil: set this to true or false
+  core-js: set this to true or false
+  core-js-pure: set this to true or false
+  cypress: set this to true or false
+  es5-ext: set this to true or false
+  esbuild: set this to true or false
+  oniguruma: set this to true or false
+  prisma: set this to true or false
+  protobufjs: set this to true or false
+  sharp: set this to true or false
+  typesense-instantsearch-adapter: set this to true or false
+  unrs-resolver: set this to true or false
+  utf-8-validate: set this to true or false
+  workerd: set this to true or false
