+++ src/pages/newsletters/ai-dev-essentials.tsx
@@ -5,7 +5,7 @@ import {useRouter} from 'next/router'
 import Image from 'next/image'
 import Link from 'next/link'
 import {getAIDevEssentialsPosts} from '@/lib/ai-dev-essentials'
-import {Post} from '@/schemas/post'
+import {Post} from '@/schemas/content/post'
 import {truncate} from 'lodash'
 import removeMarkdown from 'remove-markdown'
 import NewsletterSignupForm from '@/components/forms/newsletter-signup'
